float() added and art.py update

+++ Day 10 - Beginner - Functions with Outputs/09 - Calculator Finishing Touches and Bug Fixes/main.py
@@ -22,14 +22,14 @@ operations = {
 }
 
 def calculator():
-    num1 = int(input("What's the first number?: "))
+    num1 = float(input("What's the first number?: "))
     for symbol in operations:
         print(symbol)
     should_continue = True
 
     while should_continue == True:
         operations_symbol = input("Pick an operation: ")
-        num2 = int(input("What's the next number?: "))
+        num2 = float(input("What's the next number?: "))
 
         calculator_function = operations[operations_symbol]
         answer = calculator_function(num1, num2)
